requirements.platform.txt
fastapi
uvicorn
docker
confluent-kafka
cassandra-driver
requests